requirements.txt
feedparser>=6.0
beautifulsoup4>=4.12
trafilatura>=1.6
lxml_html_clean>=0.4
apscheduler>=3.10,<4.0
requests>=2.31
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.34
motor>=3.6
pydantic-settings>=2.7
httpx>=0.28
python-telegram-bot>=21
